+++ earthfile2llb/shell.go
@@ -37,7 +37,7 @@ func withShell(args []string, withShell bool) []string {
 	return args
 }
 
-func strWithEnvVarsAndDocker(args []string, envVars []string, withShell, withDebugger, forceDebugger, withDocker bool, exitCodeFile string, outputFile string) string {
+func strWithEnvVarsAndDocker(args []string, envVars []string, withShell, withDebugger, forceDebugger, withDocker, isExpression bool, exitCodeFile, outputFile string) string {
 	var cmdParts []string
 	cmdParts = append(cmdParts, strings.Join(envVars, " "))
 	if withDocker {
@@ -52,13 +52,15 @@ func strWithEnvVarsAndDocker(args []string, envVars []string, withShell, withDeb
 	}
 	if withShell {
 		var escapedArgs []string
+		if outputFile != "" {
+			escapedArgs = append(escapedArgs,
+				fmt.Sprintf("exec 1<>'\"'\"%s\"'\"' && ", escapeShellSingleQuotes(outputFile)))
+		}
+		if isExpression {
+			escapedArgs = append(escapedArgs, "echo")
+		}
 		for _, arg := range args {
 			escapedArgs = append(escapedArgs, escapeShellSingleQuotes(arg))
-		}
-		if outputFile != "" {
-			escapedArgs = append([]string{"echo"}, escapedArgs...)
-			escapedArgs = append(escapedArgs,
-				fmt.Sprintf(">'\"'\"%s\"'\"'", escapeShellSingleQuotes(outputFile)))
 		}
 		if exitCodeFile != "" {
 			escapedArgs = append(escapedArgs,
@@ -77,7 +79,7 @@ type shellWrapFun func(args []string, envVars []string, withShell, withDebugger,
 func withShellAndEnvVars(args []string, envVars []string, withShell, withDebugger, forceDebugger bool) []string {
 	return []string{
 		"/bin/sh", "-c",
-		strWithEnvVarsAndDocker(args, envVars, withShell, withDebugger, forceDebugger, false, "", ""),
+		strWithEnvVarsAndDocker(args, envVars, withShell, withDebugger, forceDebugger, false, false, "", ""),
 	}
 }
 
@@ -88,7 +90,7 @@ func withShellAndEnvVarsExitCode(exitCodeFile string) shellWrapFun {
 		}
 		return []string{
 			"/bin/sh", "-c",
-			strWithEnvVarsAndDocker(args, envVars, true, withDebugger, false, false, exitCodeFile, ""),
+			strWithEnvVarsAndDocker(args, envVars, true, withDebugger, false, false, false, exitCodeFile, ""),
 		}
 	}
 }
@@ -100,7 +102,19 @@ func withShellAndEnvVarsOutput(outputFile string) shellWrapFun {
 		}
 		return []string{
 			"/bin/sh", "-c",
-			strWithEnvVarsAndDocker(args, envVars, true, withDebugger, false, false, "", outputFile),
+			strWithEnvVarsAndDocker(args, envVars, true, withDebugger, false, false, false, "", outputFile),
+		}
+	}
+}
+
+func expressionWithShellAndEnvVarsOutput(outputFile string) shellWrapFun {
+	return func(args []string, envVars []string, withShell, withDebugger, forceDebugger bool) []string {
+		if !withShell {
+			panic("unexpected exec mode")
+		}
+		return []string{
+			"/bin/sh", "-c",
+			strWithEnvVarsAndDocker(args, envVars, true, withDebugger, false, false, true, "", outputFile),
 		}
 	}
 }